Add HTML endpoints to the API
It’s now possible to call (some of) the API with a browser! This is fun,
I’m gonna make this the main client instead of working on making a
javascript/purescript client.

The one downside is needing the client to refresh, but I can maybe add a
small bit of javascript later to automate that.