Remove API.APIResponse type
With the HTML repsonses, which I’m focusing on for now, it makes more
sense to embed the actions within the HTML for each page. E.g., instead
of having all actions always go in a top-level nav, it’s nicer to be
able to provide some written context and locate the form differently for
the different endpoints.